requirements.txt
asgiref==3.6.0
gunicorn
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pypi==2.1
django-environ
environ==1.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2020.1
sqlparse==0.4.3
django-cors-headers
certifi==2022.12.7
charset-normalizer==2.0.12
django-solo==2.0.0
et-xmlfile==1.1.0
idna==3.4
image==1.5.33
joblib==1.2.0
numpy==1.24.2
openpyxl==3.1.2
pandas==1.5.3
Pillow==9.4.0
# psycopg2==2.9.5
python-dateutil==2.8.2
rawkit==0.6.0
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.1.0
urllib3==1.26.14
django-crontab==0.7.1
# pycurl
